move frostologer class to own package
